Seperate Formulas from MmrService
